* Working on testing internal header serialization.

+++ GraphicalTestServer/EntryPoint.cpp
@@ -74,7 +74,7 @@ int main()
 				PlayerCreateStruct playerCreateS;
 				incomingPacket->Deserialize(playerCreateS.firstByte, playerCreateS.m_xPos, playerCreateS.m_yPos, playerCreateS.m_id, playerCreateS.name);
 
-				Packet playerCreatePacket(PacketPriority::RELIABLE_UDP);
+				Packet playerCreatePacket((int)PacketPriority::RELIABLE_UDP);
 				playerCreatePacket.Serialize(playerCreateS.firstByte, playerCreateS.m_xPos, playerCreateS.m_yPos, playerCreateS.m_id, playerCreateS.name);
 
 				PlayerCreateStruct testdelete;
@@ -100,7 +100,7 @@ int main()
 					anotherPlayerCreateS.m_id = allPlayers[i].m_id;
 			
 					strcpy_s(anotherPlayerCreateS.name, allPlayers[i].name);
-					Packet testPacket(PacketPriority::RELIABLE_UDP);
+					Packet testPacket((int)PacketPriority::RELIABLE_UDP);
 					testPacket.Serialize(anotherPlayerCreateS.firstByte, anotherPlayerCreateS.m_xPos, anotherPlayerCreateS.m_yPos, anotherPlayerCreateS.m_id, anotherPlayerCreateS.name);
 					testPeer.UDPSendTo(testPacket, client.m_ipAddress, client.m_port);
 					std::cout << "told client of another person." << std::endl;
@@ -134,7 +134,7 @@ int main()
 					{
 						allPlayers[i].m_xPos = playerMoveS.m_xPos;
 						allPlayers[i].m_yPos = playerMoveS.m_yPos;
-						Packet playerMovePacket(PacketPriority::UNRELIABLE_UDP);
+						Packet playerMovePacket((int)PacketPriority::UNRELIABLE_UDP);
 						playerMovePacket.Serialize(playerMoveS.firstByte, playerMoveS.m_xPos, playerMoveS.m_yPos, playerMoveS.m_id);
 						testPeer.UDPSendToAll(playerMovePacket);
